Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
scipy==1.11.4
statsmodels==0.14.1
weasyprint==60.1
jinja2==3.1.2